Crud clientes e criação de exception handler

+++ src/main/java/com/gabrielyorlando/locacao/repositories/LocacaoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LocacaoRepository extends JpaRepository<Locacao, Integer> {
+public interface LocacaoRepository extends JpaRepository<Locacao, Long> {
 }
